Changes to make structs refenece types, fix for increment/decrement code.

+++ src/package/oolong-module.h
@@ -1,10 +1,12 @@
 #ifndef OOLONG_MODULE_H
 #define OOLONG_MODULE_H
 
+#include <stdint.h>
+
 struct String {
     char* value;
-    long long allocatedSize;
-    long long usedSize;
+    int64_t allocatedSize;
+    int64_t usedSize;
 };
 
 #endif
